+++ gd-core/src/tests/components/d3/TestBarChart.jsx
@@ -6,177 +6,39 @@ import { BarChart } from '../../../components/d3';
 
 const FIXTURE = {
     meta: {},
-    data: [
-        {
-            year: 1980,
-            value: 5.6
-        },
-        {
-            year: 1981,
-            value: 10.1
-        },
-        {
-            year: 1982,
-            value: 11.9
-        },
-        {
-            year: 1983,
-            value: 12.9
-        },
-        {
-            year: 1984,
-            value: 11.3
-        },
-        {
-            year: 1985,
-            value: 9.4
-        },
-        {
-            year: 1986,
-            value: 10.4
-        },
-        {
-            year: 1987,
-            value: 6.1
-        },
-        {
-            year: 1988,
-            value: 8.3
-        },
-        {
-            year: 1989,
-            value: 7
-        },
-        {
-            year: 1990,
-            value: 11.9
-        },
-        {
-            year: 1991,
-            value: 12.8
-        },
-        {
-            year: 1992,
-            value: 8.3
-        },
-        {
-            year: 1993,
-            value: 19.4
-        },
-        {
-            year: 1994,
-            value: 10
-        },
-        {
-            year: 1995,
-            value: 9.5
-        },
-        {
-            year: 1996,
-            value: 11.3
-        },
-        {
-            year: 1997,
-            value: 8
-        },
-        {
-            year: 1998,
-            value: 14.4
-        },
-        {
-            year: 1999,
-            value: 11.7
-        },
-        {
-            year: 2000,
-            value: 6.4
-        },
-        {
-            year: 2001,
-            value: 8.9
-        },
-        {
-            year: 2002,
-            value: 15
-        },
-        {
-            year: 2003,
-            value: 4.6
-        },
-        {
-            year: 2004,
-            value: 10.8
-        },
-        {
-            year: 2005,
-            value: 8.2
-        },
-        {
-            year: 2006,
-            value: 6.7
-        },
-        {
-            year: 2007,
-            value: 9.2
-        },
-        {
-            year: 2008,
-            value: 7.5
-        },
-        {
-            year: 2009,
-            value: 13.2
-        },
-        {
-            year: 2010,
-            value: 12.2
-        },
-        {
-            year: 2011,
-            value: 11.7
-        },
-        {
-            year: 2012,
-            value: 4.8
-        },
-        {
-            year: 2013,
-            value: 14.5
-        },
-        {
-            year: 2014,
-            value: 6.5
-        },
-        {
-            year: 2015,
-            value: 10.1
-        },
-        {
-            year: 2016,
-            value: 12.4
-        },
-        {
-            year: 2017,
-            value: 13.2
-        }
-    ] };
+    data: []
+};
+
+for (let x = 1980; x < 2020; x += 1) {
+    const y = Math.random() * 10;
+    FIXTURE.data.push({
+        x,
+        y,
+        y0: y + (y * 0.1 * Math.random()),
+        y1: y - (y * 0.1 * Math.random())
+    });
+}
 
 const INITIAL_YEAR = 2000;
 
 const increase = (data) => (
-    data.map(({ year, value, selected }) => ({
-        year,
-        value: value + 1,
+    data.map(({ x, y, y0, y1, selected }) => ({
+        x,
+        y: y + 1,
+        y0: y0 + 1,
+        y1: y1 + 1,
         selected
     }))
 );
 
 const updateSelected = (data, selectedYear) => (
     data.map(
-        ({ year, value }) => ({
-            year,
-            value,
-            selected: year === selectedYear
+        ({ x, y, y0, y1 }) => ({
+            x,
+            y,
+            y0,
+            y1,
+            selected: x === selectedYear
         })
     )
 );
@@ -225,12 +87,10 @@ const TestBarChart = () => {
                 marginLeft={60}
                 marginRight={20}
                 xAxisProps={{
-                    key: 'year',
                     title: 'Year',
                     titlePadding: 50
                 }}
                 yAxisProps={{
-                    key: 'value',
                     title: 'lb/acre',
                     titlePadding: 30
                 }}
@@ -241,7 +101,9 @@ const TestBarChart = () => {
                 mouseOut={(d, idx, rects) => {
                     select(rects[idx]).attr('fill', '#4682b4');
                 }}
-                data={data}
+                barsData={data}
+                lineData={data}
+                intervalData={data}
             />
         </div>
     );