+++ gd-core/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - `extent` parameters can be passed to ol/Map to set bounds for the map view.
   [GEOD-1320](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1320)
+- Added support for line and confidence interval charts to BarChart
+  [GLGVO-638](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-638)
 
 ### Changed
 - Added `LegendHorizontalDiscrete` component and renamed `Legend` to `LegendHorizontalDiscrete`.