+++ gd-gltg/CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GLGVO-634](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-634)
 - Updated partners and faq pages
   [GLGVO-635](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-635)
+- Added data nitrate load to Gulf of Mexico
+  [GLGVO-638](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-638)
 
 ### Fixed
 - Click the 'x' in the top right corner of the Detail Page, the page returns to the Explore Page